// src/app/models/movie.ts
export class Movie {
   id: string;
   name: string;
   image: string;
   genre: string;
   amount: number;
   yearOfRelease: string;
   quantity: number;
   description: string;
   trailer: string;
   imdbRatings: string;
   director: string;
   cast: string;
   category: string;
   releaseDate: string
   inWishlist: boolean;
   inCart: boolean;
   Cartqty: number;
}


export class Book {
   id: string;
   name: string;
   image: string;
   genre: string;
   amount: number;
   yearOfRelease: string;
   quantity: number;
   description: string;
   preview: string;
   writer: string;
   goodreadsRatings: string;
   pages: number;
   category: string;
   releaseDate: string;
   inWishlist: boolean;
   inCart: boolean;
   Cartqty: number;
}